+++ gsl/src/gen/java/org/bytedeco/gsl/gsl_odeiv2_step.java
@@ -27,7 +27,7 @@ public class gsl_odeiv2_step extends Pointer {
         return (gsl_odeiv2_step)super.position(position);
     }
 
-  @MemberGetter public native @Const gsl_odeiv2_step_type type();
+  public native @Const gsl_odeiv2_step_type type(); public native gsl_odeiv2_step type(gsl_odeiv2_step_type setter);
   public native @Cast("size_t") long dimension(); public native gsl_odeiv2_step dimension(long setter);
   public native Pointer state(); public native gsl_odeiv2_step state(Pointer setter);
 }